* integration tests * More informative error messages * If video recorder provider is libav, use avconv binary instead of ffmpeg * Rewritten Xvfb launch using Process.spawn and avoiding a shell * Do not manually remove X11 lock file when stopping Xvfb; this isn’t conventional. Should eliminate some errors with not being able to find Xvfb * Detect situation when Xvfb can’t listen to any sockets and raise corresponding error * Properly working integration tests

+++ lib/headless/cli_util.rb
@@ -53,11 +53,13 @@ class Headless
           # Process.wait tried to wait on a dead process
         end
       end
-      
-      begin
-        FileUtils.rm pid_filename
-      rescue Errno::ENOENT
-        # pid file already removed
+
+      unless options[:preserve_pid_file]
+        begin
+          FileUtils.rm pid_filename
+        rescue Errno::ENOENT
+          # pid file already removed
+        end
       end
     end
   end
